move formatKeymap to prosemirror.ts
