+++ packages/hash/shared/src/prosemirror.ts
@@ -1,4 +1,6 @@
+import { toggleMark } from "prosemirror-commands";
 import { NodeSpec, NodeType, ProsemirrorNode, Schema } from "prosemirror-model";
+import { keymap } from "prosemirror-keymap";
 import { paragraphBlockComponentId } from "./blocks";
 
 type NodeWithAttrs<Attrs extends {}> = Omit<
@@ -354,3 +356,29 @@ export const mutateSchema = (
 export const isParagraphNode = (node: ProsemirrorNode<Schema>) => {
   return componentNodeToId(node) === paragraphBlockComponentId;
 };
+
+export const formatKeymap = (schema: Schema) =>
+  keymap<Schema>({
+    // Mod- stands for Cmd- o macOS and Ctrl- elsewhere
+    "Mod-b": toggleMark(schema.marks.strong!),
+    "Mod-i": toggleMark(schema.marks.em!),
+    "Mod-u": toggleMark(schema.marks.underlined!),
+    // We add an extra shortcut on macOS to mimic raw Chrome’s contentEditable.
+    // ProseMirror normalizes keys, so we don’t get two self-cancelling handlers.
+    "Ctrl-u": toggleMark(schema.marks.underlined!),
+
+    "Shift-Enter": (state, dispatch) => {
+      dispatch?.(
+        state.tr
+          .replaceSelectionWith(schema.nodes.hardBreak!.create())
+          .scrollIntoView(),
+      );
+      return true;
+    },
+    // execCommand is flagged as depecrated but it seems that there isn't a viable alternative
+    // to call the undo and redo default browser actions (https://stackoverflow.com/a/70831583)
+    // After the collab rework this should be replaced with a proper implementation
+    "Mod-z": () => document.execCommand("undo"),
+    "Mod-y": () => document.execCommand("redo"),
+    "Mod-Shift-z": () => document.execCommand("redo"),
+  });
